alforno_back/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.1
idna==2.10
mysqlclient==2.0.3
Pillow==8.1.2
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
stripe==2.55.2
urllib3==1.26.3
